+++ changelog.md
@@ -1,3 +1,27 @@
+Unreleased
+==========
+
+**Key breaking changes:**
+
+New features:
+
+* mask_tokens config (default: true) to mask tokens in logs
+
+Updated features:
+
+* Log before and after sending request
+
+Removed features:
+
+Internal improvements:
+
+Testing improvements:
+
+* Run Travis on latest rubies and fix config
+
+Others:
+
+
 v3.0.0 (2017-03-17)
 ======